refactor: extract ScreenCapturePermissionStatus.swift from PermissionAndPreflightTypes.swift
Splits the `ScreenCapturePermissionStatus` raw-string enum out.

Byte-identical move of lines 49-56. Both files import Foundation.

PermissionAndPreflightTypes.swift: 83 → 74 lines. Tests: 667.
