+++ Sources/BugNarrator/Services/ScreenCapturePermissionStatus.swift
@@ -1,0 +1,10 @@
+import Foundation
+
+enum ScreenCapturePermissionStatus: String, Equatable {
+    case notDetermined
+    case granted
+    case denied
+    case unavailable
+    case captureSetupFailed
+    case unknownError
+}
